net: ti: am65-cpsw-nuss: Workaround for buggy PHY/Board
On Beagleplay the PHY is not found on first attempt.
This might be due to incorrect bootstrap configuration.

The Realtec PHY treats address 0x0 as a special broadcast
address. This issue is not seen on the other PHY that
uses address 0x1.

Instead of giving up on first go we try a couple of times.

Signed-off-by: Roger Quadros <rogerq@kernel.org>